update: 1차 코드리뷰 확인 및 적용 - 변수명 수정 - NumberButton의 value 값에 Number() 사용하는 위치 변경 - index.js/ App.js 파일 분할 - 사용하지 않는 value 속성 삭제
